+++ dash_mantine_components/__init__.py
@@ -36,7 +36,7 @@ _current_path = _os.path.dirname(_os.path.abspath(__file__))
 _this_module = _sys.modules[__name__]
 
 
-async_resources = ["AreaChart", "BarChart", "LineChart", "BubbleChart", "DonutChart"]
+async_resources = ["AreaChart", "BarChart", "LineChart", "BubbleChart", "DonutChart", "PieChart"]
 
 
 _js_dist = []
